# Projects/LBAW/lbaw24023/a7_openapi.yaml
openapi: 3.0.0
info:
  version: '1.0'
  title: 'SyncTask Web API'
  description: 'Web Resources Specification (A7) for SyncTask'
servers:
- url: http://localhost:8000
  description: Development server
...
paths:
  /login:
    get:
      operationId: R101
      summary: 'R101: Login Form'
      description: 'Provide the login form for users. Access: PUB'
      tags:
        - 'M01: Authentication and Profile Management'
      responses:
        '200':
          description: 'The login UI is displayed successfully.'
        '404':
          description: 'The login page could not be found.'
        '500':
          description: 'Internal server error. Could not display the login form.'


    post:
      operationId: R102
      summary: 'R102: Login Action'
      description: 'Authenticate user credentials and establish a session. Access: PUB'
      tags:
        - 'M01: Authentication and Profile Management'
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                usernameOrEmail:
                  type: string
                  description: 'The username or email of the user.'
                password:
                  type: string
                  description: 'The password of the user.'
              required:
                - usernameOrEmail
                - password
      responses:
        '302':
          description: 'Redirect after processing login.'
          headers:
            Location:
              schema:
                type: string
              examples:
                302Success:
                  description: 'Successful login. Redirect to user dashboard.'
                  value: '/projects'
                302Failure:
                  description: 'Failed login. Redirect to login form.'
                  value: '/login'
        '401':
          description: 'Unauthorized. Incorrect username/email or password.'
        '404':
          description: 'Login endpoint not found.'
        '500':
          description: 'Internal server error during login process.'
...